feat: fixed husky
