+++ package.json
@@ -28,7 +28,7 @@
     "@astrojs/react": "3.3.0",
     "@types/react": "18.2.79",
     "@types/react-dom": "18.2.25",
-    "astro": "4.6.3",
+    "astro": "^4.6.3",
     "astrojs-service-worker": "2.0.0",
     "eslint-config-standard": "17.1.0",
     "react": "18.2.0",
