+++ src/ui/components/atoms/Closebutton/styles.scss
@@ -1,40 +1,28 @@
 @import '../../../../globals/variables';
 @import '../../../../globals/mixins';
 
-.button {
-  background: $ads-primary-main;
+.Close_button {
+  background: $ads-text-main;
   border: 2px solid transparent;
-  border-radius: 1.5rem;
   padding: 1.26rem;
   width: 100%;
 
   @include respond-to('medium') {
-    width: 80%;
+    width: 15%;
   }
 
   @include respond-to('large') {
-    width: 30%;
+    width: 15%;
   }
 
   @include respond-to('extra-large') {
-    width: 30%;
+    width: 15%;
   }
-}
-
-.button:hover {
-  background: $ads-secondary-main;
-}
-
-.button__label {
-  color: $ads-color-main;
-  font-size: $ads-size-main;
-  margin-left: 1rem;
-  margin-right: 1rem;
 }
 
 .button__icon--before {
   content: '';
-  display: inline-block;
+  display: center;
   height: 16px;
   width: 16px;
 }
