+++ src/ui/components/atoms/Closebutton/types/IProps.ts
@@ -1,7 +1,4 @@
-import React from 'react';
-
 export interface IProps {
-  children: React.ReactNode;
   variant?: string;
   icon?: string;
 }
